chore: move mode to internal
