+++ warehouse/app_test.go
@@ -3,13 +3,12 @@ package warehouse
 import (
 	"context"
 	"fmt"
+	"github.com/rudderlabs/rudder-server/warehouse/internal/mode"
 	"net"
 	"net/http"
 	"strconv"
 	"testing"
 	"time"
-
-	"github.com/rudderlabs/rudder-server/warehouse/mode"
 
 	"github.com/ory/dockertest/v3"
